make methods to create exec_context accessible to srw module
